requirements.txt
--extra-index-url https://download.pytorch.org/whl/rocm7.1
torch==2.10.0+rocm7.1
torchaudio==2.10.0+rocm7.1
transformers
datasets
librosa
soundfile
numpy
scikit-learn
tqdm
flask
flask-cors
flask-sock
aiortc
av
redis
requests
sentencepiece
accelerate
piper-tts
faiss-cpu
edge-tts
pathvalidate
gunicorn
gevent
gevent-websocket
